removed signal handler. Instead I hardcoded the function to end the program with user input

+++ src/main.cpp
@@ -7,23 +7,12 @@
 #include <thread> // this enables multiple connections
 #include <signal.h>
 #include <atomic>
-std::atomic<bool> running(true);
-
-void signalHandler(int signal)
-{
-    cout << "caught a signal " << signal << endl;
-    running = false;
-}
+#include <fcntl.h>
 
 int main()
 {
     vector<thread> threads; // to store all individual threads (sockets)
     string mode;
-    struct sigaction sigIntHandler; // handles Ctrl+C, so it doesn't unexpectedly halt the program
-    sigIntHandler.sa_handler = signalHandler;
-    sigemptyset(&sigIntHandler.sa_mask);
-    sigIntHandler.sa_flags = 0;
-    sigaction(SIGINT, &sigIntHandler, NULL);
 
     cout << "MODE: ";
     getline(cin, mode);
@@ -31,20 +20,28 @@ int main()
     if (mode == "server")
     {
         server testServer = server();
+        string command = "";
+
         testServer.establish();
         testServer.listen();
-        while (running)
+        while (true)
         {
             int sd = testServer.accept();
-            threads.emplace_back(&server::receive, &testServer, sd); // cannot read the first message for some reason
+            int recipient = testServer.selectClient(sd);
+            threads.emplace_back(&server::handleComm, &testServer, sd, recipient); // cannot read the first message for some reason
+            cout << "do you wish to disconnect the server? (Press yes to disconnect)";
+            getline(cin, command);
+            if (command == "yes")
+            {
+                for (auto &thr : threads)
+                {
+                    thr.join();
+                    cout << "thread closed" << endl;
+                }
+                cout << "Server disconnected" << endl;
+                testServer.terminate();
+            }
         }
-        for (auto &thr : threads)
-        {
-            thr.join();
-            cout << "thread closed" << endl;
-        }
-        cout << "Server disconnected" << endl;
-        testServer.terminate();
     }
     else
     {
@@ -52,14 +49,22 @@ int main()
         testClient.connect();
         char message[1024];
 
-        while (running)
+        string command;
+
+        while (true)
         {
+            testClient.receive();
             cout << "message: ";
             fgets(message, sizeof(message), stdin);
             cout << "sending..." << endl;
             testClient.send(message);
+            cout << "do you wish to disconnect? Press yes to disconnect";
+            getline(cin, command);
+            if (command == "yes")
+            {
+                testClient.terminate();
+            }
         }
-        testClient.terminate();
     }
     return 0;
 }
